[CLN] log-service: move scrub to its own file
